+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- 2018.12.04, 10:40 -- LoadingSpinner on prefetch of Gallery images
+- 2018.12.04, 10:40 -- `LoadingSpinner` splash while loading.
 - 2018.12.03, 08:01 -- Markdown render: custom comments, code blocks (highlight.js), html parser.
 - 2018.12.01, 23:21 -- Animate rendered content (`Error` or `LoadedPage`) in `LoadPage`? +
 - 2018.11.29, 08:29 -- Real page sources fetching.